LPS-108858 Refactor: Move classes to model package
