+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/java/com/liferay/analytics/reports/web/internal/model/TimeRange.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.analytics.reports.web.internal.data.time;
+package com.liferay.analytics.reports.web.internal.model;
 
 import java.time.Clock;
 import java.time.LocalDate;
