+++ modules/dxp/apps/analytics/analytics-reports-web/src/test/java/com/liferay/analytics/reports/web/internal/model/TimeRangeTest.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.analytics.reports.web.internal.data.time;
+package com.liferay.analytics.reports.web.internal.model;
 
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
